Sistema de login y permisos

+++ procedimientos.php
@@ -1,4 +1,5 @@
 <?php
+session_start();
 //Configuración Global
 
 	//Método conexión 2
@@ -13,13 +14,32 @@
         $ssh->login('pirate', $key);
 
         $salida = array_slice($_POST, 0);
-
-    foreach($salida as $linea){
-        echo "<p>".$linea."</p>";
-        $ssh->exec("docker stop ".$linea." ");
-        $ssh->exec("docker rm ".$linea." ");	
-        $ssh->exec("docker volume rm ".$linea." ");	
-    }
+        
+        if ($_POST["Accion"] == "Borrar"){
+            unset($_POST["Accion"]);
+            foreach($salida as $linea){
+                echo "<p>".$linea."</p>";
+                $ssh->exec("docker stop ".$linea." ");
+                $ssh->exec("docker rm ".$linea." ");	
+                $ssh->exec("docker volume rm ".$linea." ");	
+            }
+        }
+        elseif ($_POST["Accion"] == "Parar"){
+            unset($_POST["Accion"]);
+            foreach($salida as $linea){
+                echo "<p>".$linea."</p>";
+                $ssh->exec("docker stop ".$linea." ");
+            }
+        }
+        elseif ($_POST["Accion"] == "Start"){
+            unset($_POST["Accion"]);
+            foreach($salida as $linea){
+                echo "<p>".$linea."</p>";
+                $ssh->exec("docker start ".$linea." ");
+            } 
+        }
+    
     header("Location: http://85.136.104.237/index.php");
     exit();
+
 ?>
